moved some required files to public headers
