+++ src/utils/per_index_bitset.cpp
@@ -1,4 +1,4 @@
-#include "per_index_bitset.h"
+#include "../../include/dlplan/per_index_bitset.h"
 
 /**
  * Taken from fast-downward.org
